added layout for EmployeesScreen.kt and new SearchLikeScreen.kt, created new ui components for it,

+++ core/networkApi/employees/api/src/main/java/odoo/miem/android/core/networkApi/employees/api/source/ResponseEntities.kt
@@ -17,7 +17,7 @@ data class AllEmployeesResponse(
 
         @Json(name = "id")
         @SpecifiedTypeOrNull(JsonReader.Token.NUMBER)
-        val id: Int?,
+        val id: Long?,
 
         @Json(name = "name")
         @SpecifiedTypeOrNull(JsonReader.Token.STRING)
@@ -46,7 +46,7 @@ data class EmployeeDetailsResponse(
 
     @Json(name = "id")
     @SpecifiedTypeOrNull(JsonReader.Token.NUMBER)
-    val id: Int?,
+    val id: Long?,
 
     @Json(name = "name")
     @SpecifiedTypeOrNull(JsonReader.Token.STRING)
